requirements.txt
kagglehub
numpy 
pandas 
matplotlib
seaborn
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-chroma
torch
sentence-transformers
gradio
notebook
ipywidgets